requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
h11==0.14.0
mysqlclient==2.2.4
packaging==24.0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0
